voice-training/pyproject.toml
[project]
name = "yesman-voice-training"
version = "0.1.0"
description = "Yes Man音声クローニング学習環境"
requires-python = ">=3.11"
dependencies = [
    "librosa>=0.10.1",
    "soundfile>=0.12.1",
    "pydub>=0.25.1",
    "moviepy>=2.0.0",
    "numpy==1.26.4",
    "tqdm>=4.66.1",
    "pyyaml>=6.0.1",
    "ffmpeg-python>=0.2.0",
    "TTS>=0.22.0",
    "openai-whisper>=20231117",
    "torch>=2.1.0",
    "torchaudio>=2.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["scripts", "tts"]
